explain erupt-flow comment
