+++ erupt-data/erupt-mybatis-query/pom.xml
@@ -3,13 +3,13 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>erupt-mybatis</artifactId>
-    <description>EruptProcess Mybatis impl</description>
+    <artifactId>erupt-mybatis-query</artifactId>
+    <description>Erupt Mybatis query</description>
 
     <parent>
         <groupId>xyz.erupt</groupId>
         <artifactId>erupt</artifactId>
-        <version>1.10.8</version>
+        <version>1.12.17</version>
         <relativePath>../../pom.xml</relativePath>
     </parent>
 
@@ -20,9 +20,9 @@
             <version>${project.parent.version}</version>
         </dependency>
         <dependency>
-            <groupId>com.baomidou</groupId>
-            <artifactId>mybatis-plus-boot-starter</artifactId>
-            <version>3.4.1</version>
+            <groupId>org.mybatis.spring.boot</groupId>
+            <artifactId>mybatis-spring-boot-starter</artifactId>
+            <version>2.3.2</version>
         </dependency>
     </dependencies>
