+++ erupt-data/erupt-mybatis-query/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,1 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=xyz.erupt.mybatis.EruptMybatisAutoConfiguration
