+++ erupt-data/erupt-mybatis-query/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
