--- erupt-data/erupt-mybatis/README.md
@@ -1,1 +1,0 @@
-### erupt Mybatis ORM
