--- erupt-data/erupt-mybatis/src/main/resources/META-INF/spring.factories
@@ -1,1 +1,0 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=xyz.erupt.es.EruptMybatisAutoConfiguration
