+++ erupt-extra/erupt-flow/src/main/java/xyz/erupt/flow/controller/EruptFlowFileController.java
@@ -31,8 +31,6 @@ public class EruptFlowFileController {
 
     /**
      * 上传图片
-     * @param file
-     * @return
      */
     @SneakyThrows
     @PostMapping("/upload")
