+++ erupt-extra/erupt-flow/src/main/java/xyz/erupt/flow/process/listener/ExecutableNodeListener.java
@@ -9,7 +9,6 @@ public interface ExecutableNodeListener<T> extends Comparable<ExecutableNodeList
 
     /**
      * 默认顺序，最大
-     * @return
      */
     default int sort() {
         return Integer.MIN_VALUE;
