+++ erupt-extra/erupt-flow/src/main/java/xyz/erupt/flow/repository/EruptOrgRepository.java
@@ -9,14 +9,11 @@ public interface EruptOrgRepository extends JpaRepository<EruptOrg, Long> {
 
     /**
      * 查询上级为空的单位
-     * @return
      */
     List<EruptOrg> findByParentOrgIdIsNullOrderBySortAsc();
 
     /**
      * 根据上级机构id查询
-     * @param orgId
-     * @return
      */
     List<EruptOrg> findByParentOrgIdOrderBySortAsc(Long orgId);
 }
