+++ erupt-extra/erupt-flow/src/main/java/xyz/erupt/flow/service/TaskHistoryService.java
@@ -11,22 +11,13 @@ public interface TaskHistoryService extends IService<OaTaskHistory> {
 
     /**
      * 拷贝并保存
-     * @param task
-     * @return
      */
     public OaTaskHistory copyAndSave(OaTask task);
 
     /**
      * 拷贝并保存
-     * @param tasks
-     * @return
      */
     public List<OaTaskHistory> copyAndSave(Collection<OaTask> tasks);
 
-    /**
-     *
-     * @param id
-     * @return
-     */
     List<OaTaskHistory> listByActivityId(Long id);
 }
