+++ pom.xml
@@ -61,7 +61,6 @@
         <!--使用H2作为数据源的示例项目-->
         <module>erupt-sample</module>
 
-
         <!--erupt deploy-->
         <module>deploy/erupt-cloud-server-docker</module>
     </modules>
